Addressing comments from PR and some refactor
