+++ app/templates/dashboard.html
@@ -3,59 +3,47 @@
 {%- from "govuk_frontend_jinja/components/button/macro.html" import govukButton -%}
 {% from "partials/start-application-button.html" import startApplicationButton %}
 {% from "partials/available_in_language.html" import available_in_language %}
+{% from "partials/round_closed_warning.html" import round_closed_warning %}
 
 {% extends "base.html" %}
 {% block content %}
-    <h1 class="govuk-heading-xl">{% trans %}Your applications{% endtrans %}</h1>
-{% if submission_deadline|current_datetime_after_given %}
-        <div class="govuk-notification-banner" role="region" aria-labelledby="govuk-notification-banner-title" data-module="govuk-notification-banner">
-        <div class="govuk-notification-banner__header">
-          <h2 class="govuk-notification-banner__title" id="govuk-notification-banner-title">
-            Important
-          </h2>
-        </div>
-        <div class="govuk-notification-banner__content">
-          <h2 class="govuk-notification-banner__heading">
-            Round closed - {{fund_name}} {{ round_title }}
-          </h2>
-          <p class="govuk-body">(Closed: {{ submission_deadline|datetime_format }})</p>
-          <p class="govuk-body">You can no longer submit applications in this window. Any you had in progress will be emailed to you.</p>
-          <p class="govuk-body">We'll announce the next application window dates shortly.</p>
-        </div>
-      </div>
-        {% endif %}
-    <p class="govuk-body">
-        {% trans %}You have started{% endtrans %}&nbsp;{% trans count=applications|length %}{{ count }} application{% pluralize %}{{ count }} applications{% endtrans %}&nbsp;{% trans %}using this email address{% endtrans %}.
-    </p>
-{{available_in_language(inset=False, warn=True)}}
-    {% set ns = namespace(rows = []) %}
-    {% for application in applications %}
-        {% set status_class %}
-            {% if submission_deadline|current_datetime_after_given  %}
-                {%  if application.status == 'SUBMITTED'%}
-                    complete-tag
-                {%  elif application.status == 'WINDOW_CLOSED' %}
-                    complete-tag
-                {% endif %}
-            {% endif %}
-            {% if submission_deadline|current_datetime_before_given  %}
-                {%  if application.status == 'SUBMITTED'%}
-                    complete-tag
-                {%  elif application.status == 'NOT_STARTED' %}
-                    not-started-tag
-                {%  elif application.status == 'IN_PROGRESS' %}
-                    in-progress-tag-new
-                {%  elif application.status == 'READY_TO_SUBMIT' %}
-                    in-progress-tag-new
-                {% endif %}
-            {% endif %}
-        {% endset %}
+<h1 class="govuk-heading-xl">{% trans %}Your applications{% endtrans %}</h1>
+{% if is_past_submission_deadline %}
+    {{ round_closed_warning(fund_name, round_title, submission_deadline) }}
+{% endif %}
+<p class="govuk-body">
+    {% trans %}You have started{% endtrans %}&nbsp;{% trans count=applications|length %}{{ count }} application
+    {% pluralize %}{{ count }} applications{% endtrans %}&nbsp;{% trans %}using this email address{% endtrans %}.
+</p>
+{{ available_in_language(inset=False, warn=True) }}
+{% set ns = namespace(rows = []) %}
+{% for application in applications %}
+{% set status_class %}
+{% if is_past_submission_deadline %}
+    {% if application.status == 'SUBMITTED' %}
+        complete-tag
+    {% elif application.status == 'WINDOW_CLOSED' %}
+        complete-tag
+    {% endif %}
+{% endif %}
+{% if not is_past_submission_deadline %}
+{% if application.status == 'SUBMITTED' %}
+    complete-tag
+{% elif application.status == 'NOT_STARTED' %}
+    not-started-tag
+{% elif application.status == 'IN_PROGRESS' %}
+    in-progress-tag-new
+{% elif application.status == 'READY_TO_SUBMIT' %}
+    in-progress-tag-new
+{% endif %}
+{% endif %}
+{% endset %}
 
-        {% if not application.status == 'SUBMITTED' and submission_deadline|current_datetime_before_given %}
-            {% set application_link = '<a class="govuk-link" href="{}">{}</a>'.format(url_for('routes.tasklist', application_id=application.id), gettext('Continue application')) %}
-        {% endif %}
+{% if not application.status == 'SUBMITTED' and not is_past_submission_deadline %}
+    {% set application_link = '<a class="govuk-link" href="{}">{}</a>'.format(url_for('routes.tasklist', application_id=application.id), gettext('Continue application')) %}
+{% endif %}
 
-        {% set row = ns.rows.append(
+{% set row = ns.rows.append(
         [
             {
                 'text': application.project_name if application.project_name else gettext('Untitled project')
@@ -75,12 +63,11 @@
             {
                 'html': application_link
             }
-        ])%}
-    {% endfor %}
+        ]) %}
+{% endfor %}
 
-    {% if applications|length > 0 and submission_deadline|current_datetime_before_given %}
-        {{
-            govukTable({
+{% if applications|length > 0 and not is_past_submission_deadline %}
+    {{ govukTable({
             'head': [
                 {
                     'text': gettext('Project name')
@@ -99,14 +86,12 @@
                 }
             ],
             'rows': ns.rows
-            })
-        }}
-    {% endif %}
+            }) }}
+{% endif %}
 
 
-    {% if applications|length > 0 and submission_deadline|current_datetime_after_given %}
-        {{
-            govukTable({
+{% if applications|length > 0 and is_past_submission_deadline %}
+    {{ govukTable({
             'head': [
                 {
                     'text': gettext('Project name')
@@ -122,10 +107,9 @@
                 }
             ],
             'rows': ns.rows
-            })
-        }}
-    {% endif %}
-{% if submission_deadline|current_datetime_before_given %}
-    {{startApplicationButton(form, url_for('routes.new'), applications=applications)}}
+            }) }}
+{% endif %}
+{% if not is_past_submission_deadline %}
+    {{ startApplicationButton(form, url_for('routes.new'), applications=applications) }}
 {% endif %}
 {% endblock %}
